requirements/base.txt
asgiref==3.7.2
boto3==1.28.3
botocore==1.31.3
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-ckeditor==6.5.1
django-environ==0.4.5
django-js-asset==2.0.0
django-storages==1.9.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.17.1
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==1.0.1
jsonfield==3.1.0
MarkupSafe==1.1.1
packaging==20.4
psycopg2-binary==2.9.6
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2023.3
requests==2.24.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.15.0
sqlparse==0.4.4
uritemplate==3.0.1
urllib3==1.25.11